This adds an optional validation after executing an NVFuser node, which checks that the output is the same as the unfused implementation. Then the outputs and the graph are reported via a callback.

```python
import torch

def callback(x, y, graph):
    for i in range(len(x)-amt, len(x)):
        print(x[i])
        print(y[i])
    print(graph)


with torch.jit.fuser("fuser2"):
    torch._C._jit_nvfuser_set_comparison_callback(True, callback)

    torch.jit.script
    def g(x, y):
        z = torch.add(x, y)
        return torch.sin(z)

    def f(x, y, a):
        z = torch.add(x, y)
        return g(torch.relu(z), a)

    f_s = torch.jit.script(f)
    x = torch.rand((10, 10), dtype=torch.half).cuda()
    y = torch.rand((10, 10), dtype=torch.half).cuda()
    a = torch.rand((10, 10), dtype=torch.half).cuda()
    f_s(x, y, a)
    f_s(x, y, a)
    f_s(x, y, a)
```

Differential Revision: [D34975310](https://our.internmc.facebook.com/intern/diff/D34975310)

[ghstack-poisoned]

+++ torch/csrc/jit/codegen/cuda/manager.cpp
@@ -222,7 +222,8 @@ void runCudaFusionGroup(const Node* fusion_node, Stack& stack) {
   auto compare_callback = getCudaFuserComparisonCallback();
   if (compare_callback.run_fallback) {
     // make a copy of the stack
-    int64_t inputs_size = fusion_node->g(attr::Subgraph)->inputs().size();
+    int64_t inputs_size =
+        static_cast<int64_t>(fusion_node->g(attr::Subgraph)->inputs().size());
     TORCH_INTERNAL_ASSERT(stack.size() >= inputs_size);
     stack_copy = Stack();
     stack_copy->insert(
@@ -280,7 +281,8 @@ void runCudaFusionGroup(const Node* fusion_node, Stack& stack) {
   if (compare_callback.callback != nullptr) {
     Stack fused_outputs;
     Stack fallback_outputs;
-    int64_t output_count = fusion_node->g(attr::Subgraph)->outputs().size();
+    int64_t output_count =
+        static_cast<int64_t>(fusion_node->g(attr::Subgraph)->outputs().size());
     TORCH_CHECK(
         output_count <= stack.size(),
         "Expected ",
